+++ src/main.c
@@ -6,15 +6,15 @@
 /*   By: tvogel <tvogel@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/05 21:45:04 by tvogel            #+#    #+#             */
-/*   Updated: 2022/03/11 12:14:56 by tvogel           ###   ########.fr       */
+/*   Updated: 2022/03/11 20:36:26 by tvogel           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-int	g_return;
-int	g_pid;
-int g_child;
+u_int8_t	g_return;
+int			g_pid;
+int			g_child;
 
 int	main(int argc __attribute__((unused)), char **argv __attribute__((unused)), char **envp)
 {
